Renamed tempTitle to tempOverlay and added tempTitleCountdown
