+++ Features/vexo/Dungeons/allLeapedP5.js
@@ -3,7 +3,7 @@ import { registerWhen } from "../../../../BloomCore/utils/Utils"
 import { onChatPacket } from "../../../../BloomCore/utils/Events";
 import { getAllPlayers } from '../../../utils/util';
 import { GuiEditor } from '../../../utils/GuiEditor';
-import { tempTitle } from '../../../utils/tempTitle';
+import { tempTitle } from '../../../utils/tempOverlay';
 
 let isInP5 = false;
 let notifiedallLeaped = false;
