+++ index.js
@@ -11,7 +11,8 @@ import "./Features/vexo/QOL/rejoin"
 import "./Features/vexo/Dungeons/StopMageSheepRendering"
 import "./Features/vexo/Dungeons/healerLeapedpre4"
 
-import { GUICommand } from "./Features/gui"
+
 import { RewarpCommand } from "./Features/vexo/QOL/rewarp"
 import { TYFRCommand } from "./Features/vexo/Dungeons/tyfr"
 import { SuckTrapCommand } from "./Features/vexo/Dungeons/SuckTrap"
+import { GUICommand } from "./Features/gui"
